Fix for Storag kor being colonized at level 1 with a built industry building
Bug:
Colonizing Storag Kar on the the vortex campaign instantly sets the main chain to level 1 instead of ruined and gifts the settlement a wh2_main_def_industry_1 building. The cause of the issue seems to be the start_pos_settlements table.

Solution:
Use a script to tansfer the settlement to the wh2_main_def_hag_graef faction and then set the settlement abandoned.