Fixed Empire interventions take a lord at random from your pool and KILL it after the battle is done
Bug:
Empire interventions take a lord at random from your pool and KILL it after the battle is done

Solution:
Changed to the wh2_dlc13_empire_politics.lua file had to be made:
It now saves the cqi of all lords already recruited and not leading armies or being wounded at that time. When the invasion defence force is done fighting it compares the cqi of the forces leader with the cqi table. If the leader is a already recruited lord he/she gets immortality. When the script kills the lord kills the lord and its army, a listener takes the immortality away.
That means the lord is mortal during the fight and only get immortality right before lord & force are deleted and it's taken away again right after.