Fix for Imrik Dragon encounter script error
Bug:
Imrik Dragon Encounter script throws an error if Imrik looses his last settlement and the script is supposed to spawn a new encounter.

Solution:
Use the position of the the first general of the remaining military forces as backup if both imrik is wounded and there is no region left.