Fix for: Supply line penalties take 1 turn to update
Bug:
Supply line penalties take 1 turn to update if the a force was killed or disbanded.

Solution:
Listen for CharacterConvalescedOrKilled and execute the apply_upkeep_penalty function if the killed character is a general type and delay the the execution of the apply_upkeep_penalty to make sure the recently killed force is properly dead.