Fixed by CA: Imrik Dragon encounter script error
Kind of fixed by CA. The difference between our fix and theirs is that their change results in the Dragon encounter to no longer fire if the player doesn't have a home region whereas ours just uses the position of the last army as a reference for the position of the next dragon encounter.
Good enough imo.