Fix for Confederating Throt before level 5 will cause the last quest mission to not spawn ghoritch
Bug:
From the code comment:
"ideally we would care for the edge cases where throt is confederated before this happens so ghoritch appears regardless, but at rank 5 the odds are very low"

Solution:
Take care of the edge case by changing the if clause. Also added a check for a saved_value to prevent any Ghoritch doppelganger.