Fix for warlock master engineer model bug
>Getting a bug where the warlock master engineer changes model when you uses a mount (flayer or wheel).
This happens on the campaign map he will be riding it with the normal model, but when he is in battles his unit card will be the one from the campaign map but the in battle one will be a different warlock master.
When you finish the fight he will be the same normal model again.