Fix for Welf Shadow Spellsinger missing scouting skill in vortex
Welf Shadow Spellsinger (that's the magic hero), and only Shadow (Beasts & Life are fine), is missing a place to plonk skill points for Scouting in my Vortex Sisters campaign. Some poking later, the fault is in character_skill_nodes_tables , whereby in the Campaign Key column it's specified main_warhammer , which presumably stops it from working in Vortex. Clearing that cell fixes it