Fix for Beastmen loc inconsistencies
> 1. Empire of Beasts tree.
The description of actual bonuses provided in all 4 technologies there has extra dots at the end, which is not the case in the vast majority of other races' technologies. (Example: the Gaze of the Dark Gods challenge says "Passive ability: "Regeneration" for Chaos Spawn and Jabberslythe units." That dot shouldn't be there. Same for most other bonus descriptions in that tree.)
2. Days of Wrath tree.
Same as previous, but there's also the inconsistency in the Rampage! and the Rewards of Ruin which I'd rather illustrate with two pictures attached below.
Also, the Roaming Torment tech has Cygors spelled "Cygor Units" and "Cygor units" in the two bonuses it gives. Not consistent.

And a bunch more