Fix for Belegar Ironhammer no longer receives An Oath Fulfilled when capturing Karak Eight Peaks
Bug:
Part of the Karak Eight Peaks was set up in script before the dwarf "rework". Now the whole grudge mission is set up in db, but CA forgot to add a payload and forgot to link the effect bundle to it.

>Legendary Grudges award considerably more Oathgold than generated Grudges and, in some cases, powerful buffs or unique Master Runes (see below).

Solution:
Adding missing effect bundle and oathgold payload. Will fix the issue for new campaigns, I think.