Fix for cst admiral tech is missing lores of magic information
> Vampire Coast tech tree has 4 legendary lords (admirals) it can unlock.
Each of these lords is a caster (Vampire or Deeps or Death).
Until one of the latest patches the tech tree research tooltip for each of these lords would state what Lore of magic they use but that information is now missing.
The info that is missing and needs to be added back is:
- Disreputable Admiral uses Lore of Vampires
- Infamous Admiral is Lore of Deeps
- Notorious admiral is Lore of Death
- Dread admiral is lore of Deeps
Proof of this can be found in various older screenshots and videos of VCoast gameplay.